[FIX] 스플래시뷰 @MainActor 추가(#53)
