+++ Genti_iOS/Genti_iOS/App/GentiApp.swift
@@ -11,7 +11,7 @@ struct GentiApp: View {
 
     var body: some View {
         RoutingView(Router<MainRoute>()) { router in
-            SplashView(splashViewModel: SplashViewModel(router: router, authRepository: AuthRepositoryImpl(requestService: RequestServiceImpl()), userdefaultRepository: UserDefaultsRepositoryImpl()))
+            SplashView(splashViewModel: SplashViewModel(router: router, splashUseCase: SplashUseCaseImpl(authRepository: AuthRepositoryImpl(requestService: RequestServiceImpl()), userdefaultRepository: UserDefaultsRepositoryImpl())))
                 .onNotificationRecieved(name: Notification.Name(rawValue: "PushNotificationReceived")) { _ in
                     router.routeTo(.completeMakeImage)
                     print(#fileID, #function, #line, "- noti received")
