+++ Genti_iOS/Genti_iOS/SplashView/UseCase/SplashUseCase.swift
@@ -1,0 +1,12 @@
+//
+//  SplashUseCase.swift
+//  Genti_iOS
+//
+//  Created by uiskim on 7/31/24.
+//
+
+import Foundation
+
+protocol SplashUseCase {
+    func canAutoLogin() async -> Bool
+}
